Update cart.php
In this updated code, we added a checkbox for each product in the cart, and updated the logic for calculating the total to only include the selected products. When the user clicks the "Update Cart" button, the quantities for the selected products will be updated, and the total will only reflect the selected products. Similarly, when the user clicks the "Delete Selected" button, only the selected products will be deleted from the cart and the total will be updated accordingly.